yt-video-search/requirements.txt
aiobotocore==1.1.2
aiohttp==3.7.2
aioitertools==0.7.0
altair==4.1.0
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.12.1
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
base58==2.1.0
black==20.8b1
bleach==3.2.1
blinker==1.4
botocore==1.17.44
cachetools==4.2.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
clip==1.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
fsspec==0.8.4
ftfy==6.0.1
gitdb==4.0.7
GitPython==3.1.14
great-expectations==0.12.6
hjson==3.0.2
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
mistune==0.8.4
multidict==5.0.0
mypy-extensions==0.4.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
notebook==6.1.4
numpy==1.19.2
opencv-python==4.5.1.48
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.15.8
ptyprocess==0.6.0
pyarrow==3.0.0
pycparser==2.20
pydeck==0.6.2
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
regex==2020.10.23
requests==2.23.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3fs==0.5.1
scipy==1.5.3
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
streamlit==0.80.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
toolz==0.11.1
torch==1.7.1
torchvision==0.8.2
tornado==6.0.4
tqdm==4.60.0
traitlets==5.0.5
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
validators==0.18.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.2
youtube-dl==2021.4.26
zipp==3.3.2
